requirements.txt
Blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
numpy==2.1.2
opencv-python==4.10.0.84
Werkzeug==3.0.4
scikit-image==0.24.0
imutils==0.5.4
mysql-connector-python==8.0.33
